requirements.txt
rest3client
github3api
semantic_version
ansicolors